Fix dimension id read from login packet. This was working only because of pure luck, but that explains problems with players logging in non in overworld dimension.

+++ src/protocolsupport/protocol/packet/middle/clientbound/play/MiddleLogin.java
@@ -19,7 +19,7 @@ public abstract class MiddleLogin<T> extends ClientBoundMiddlePacket<T> {
 	public void readFromServerData(PacketDataSerializer serializer) {
 		playerEntityId = serializer.readInt();
 		gamemode = serializer.readByte();
-		dimension = serializer.readByte();
+		dimension = serializer.readInt();
 		difficulty = serializer.readByte();
 		serializer.readByte();
 		maxplayers = TabAPI.getMaxTabSize();
